cambios
cambios

+++ usuario/src/main/java/com/taller/usuario/services/UsuarioService.java
@@ -4,7 +4,7 @@ import com.taller.usuario.model.entities.Usuario;
 import java.util.List;
 import java.util.Optional;
 
-public interface UsuarioService {
+public interface UsuarioService  {
     /**
      * Obtiene la lista de todos los usuarios.
      * @return Lista de usuarios.
@@ -23,11 +23,11 @@ public interface UsuarioService {
      * @param id El ID del usuario.
      * @return Un Optional con el usuario si se encuentra.
      */
-    Optional<Usuario> obtenerPorId(int id);
+    Optional<Usuario> obtenerPorId(Long id); // Eliminamos la versión con int
 
     /**
      * Elimina un usuario por su ID.
      * @param id El ID del usuario a eliminar.
      */
-    void eliminar(int id);
+    void eliminar(Long id); // Eliminamos la versión con int
 }
